[#1] refactor: View 모듈화 및 프로필카드 스택 구조 변경
