+++ C2_DiverDexApp/Views/Home/HomeView.swift
@@ -9,36 +9,15 @@ import SwiftUI
 
 struct HomeView: View {
     let profiles = ProfileCardModel.sampleProfileDatas
-    let columns = [GridItem(.flexible()), GridItem(.flexible())]
 
     var body: some View {
-
         ZStack {
             Color("Background")
+
             VStack(spacing: 15) {
-                // 상단 로고, 프로필 아이콘
-                HStack {
-                    Text("DiverDex")
-                        .font(.title)
-                        .fontWeight(.bold)
-                        .foregroundStyle(Color("KeyColor"))
-                        
-                    Spacer()
-                }
-                .padding(.horizontal)
-                .padding(.top, 15)
-
+                HeaderView()
                 SearchView()
-
-                ScrollView {
-                    LazyVGrid(columns: columns) {
-                        ForEach(profiles) { profile in
-                            ProfileCardView(model: profile)
-                                .padding(.vertical, 8)
-                        }
-                    }
-                    .padding(.horizontal)
-                }
+                ProfileGridView(profiles: profiles)
             }
         }
         .hideKeyboardWhenTappedAround()
