--- C2_DiverDexApp/Views/ProfileCardView.swift
@@ -1,60 +1,0 @@
-//
-//  ProfileCard.swift
-//  C2_DiverDexApp
-//
-//  Created by PenguinLand on 4/16/25.
-//
-
-import SwiftUI
-
-struct ProfileCardView: View {
-    let model: ProfileCardModel
-    
-
-    var body: some View {
-        RoundedRectangle(cornerRadius: 12)
-            .foregroundStyle(Color("ProfileCardBackground"))
-
-            .frame(width: 164, height: 111)
-            .shadow(color: Color(.systemGray3), radius: 2, x: 1, y: 1)
-            .overlay {
-                HStack {
-                    VStack {
-                        Text(model.name)
-                            .font(.title3)
-                            .fontWeight(.semibold)
-
-                        Capsule()
-                            .foregroundStyle(Color("ProfileCardBackground"))
-                            .opacity(0.2)
-                            .shadow(color: .black, radius: 1, x: 0.7, y: 0.5)
-                            
-                            .overlay {
-                                Text(model.category.rawValue)
-                                    .font(.caption)
-                                    .fontWeight(.semibold)
-                                    .foregroundStyle(Color(model.category.color))
-                            }
-                            .frame(width: model.category.boxSize)
-                            
-                    }
-                    .padding(.leading, 15)
-                    .padding(.top, 10)
-
-                    Spacer()
-
-                    Image(model.ProfileImageName)
-                        .resizable()
-                        .scaledToFit()
-                }
-                .frame(maxWidth: .infinity)
-                .frame(height: 70)
-                .padding(.trailing, 10)
-            }
-
-    }
-}
-
-#Preview {
-    ProfileCardView(model: ProfileCardModel.sampleProfileDatas[0])
-}
